+++ tests/ConfigFetcherTest.php
@@ -13,11 +13,37 @@ final class ConfigFetcherTest extends TestCase
 	protected function setUp(): void
 	{
 		unset($GLOBALS['debugbundle_wp_test_remote_request_response']);
+		unset($GLOBALS['debugbundle_wp_test_last_remote_request']);
 	}
 
 	protected function tearDown(): void
 	{
 		unset($GLOBALS['debugbundle_wp_test_remote_request_response']);
+		unset($GLOBALS['debugbundle_wp_test_last_remote_request']);
+	}
+
+	public function testRequestUrlMethodAndHeadersAreSanitizedBeforeRemoteFetch(): void
+	{
+		$GLOBALS['debugbundle_wp_test_remote_request_response'] = [
+			'response' => ['code' => 200],
+			'body' => '{}',
+		];
+
+		$fetcher = new ConfigFetcher();
+		$fetcher(" https://api.debugbundle.com/v1/sdk/config\n", [
+			'method' => " get<script>ignored</script>",
+			'headers' => [
+				'X-DebugBundle-Service' => ' wordpress <b>site</b> ',
+				"Bad\nHeader" => 'ignored',
+			],
+		]);
+
+		self::assertSame('https://api.debugbundle.com/v1/sdk/config', $GLOBALS['debugbundle_wp_test_last_remote_request']['url']);
+		self::assertSame('GET', $GLOBALS['debugbundle_wp_test_last_remote_request']['args']['method']);
+		self::assertSame(
+			['X-DebugBundle-Service' => 'wordpress site'],
+			$GLOBALS['debugbundle_wp_test_last_remote_request']['args']['headers']
+		);
 	}
 
 	public function testWpErrorThrowsStablePluginMessage(): void